implement basic abstract type handling and split typ::mod into multiple files by function
